requirements.txt
absl-py==0.9.0
altgraph==0.17
astor==0.8.1
backcall==0.1.0
beautifulsoup4==4.9.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cycler==0.10.0
decorator==4.4.2
gast==0.3.3
grpcio==1.29.0
gTTS==2.1.1
gTTS-token==1.1.3
h5py==2.10.0
hanziconv==0.3.2
idna==2.9
importlib-metadata==1.6.0
ipython==7.14.0
ipython-genutils==0.2.0
jedi==0.17.0
joblib==0.15.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.2.2
matplotlib==3.2.1
mock==4.0.2
numpy==1.18.4
pandas==1.0.3
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
pkg-resources==0.0.0
prompt-toolkit==3.0.5
protobuf==3.12.0
ptyprocess==0.6.0
pygame==1.9.6
Pygments==2.6.1
PyInstaller==3.6
pyparsing==2.4.7
PyQt5==5.14.2
PyQt5-sip==12.7.2
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
scikit-learn==0.23.1
scipy==1.4.1
six==1.14.0
sklearn==0.0
soupsieve==2.0.1
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
termcolor==1.1.0
threadpoolctl==2.0.0
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.1.9
Werkzeug==1.0.1
zipp==3.1.0
